requirements.txt
pymongo
certifi
alpaca-py
yfinance[nospam]
scipy
pandas
cython
fastapi
uvicorn
motor
python-dotenv
stats
PyWavelets
ripser
datetime
TA-Lib
pandas-ta
pandas_market_calendars
lxml
git+https://github.com/AmpyFin/quantstats.git
wandb
pre-commit
black==25.1.0
isort==6.0.1
flake8==7.1.2
requests_ratelimiter